+++ android/src/main/java/com/dooboolab/flutterinapppurchase/AndroidInappPurchasePlugin.java
@@ -148,7 +148,6 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
         for (Purchase purchase : purchases) {
           final ConsumeParams consumeParams = ConsumeParams.newBuilder()
               .setPurchaseToken(purchase.getPurchaseToken())
-              .setDeveloperPayload(purchase.getDeveloperPayload())
               .build();
 
           final ConsumeResponseListener listener = new ConsumeResponseListener() {
@@ -267,7 +266,6 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
             item.put("transactionReceipt", purchase.getOriginalJson());
             item.put("orderId", purchase.getOrderId());
             item.put("purchaseToken", purchase.getPurchaseToken());
-            item.put("developerPayloadAndroid", purchase.getDeveloperPayload());
             item.put("signatureAndroid", purchase.getSignature());
             item.put("purchaseStateAndroid", purchase.getPurchaseState());
 
@@ -314,7 +312,6 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
               item.put("purchaseToken", purchase.getPurchaseToken());
               item.put("dataAndroid", purchase.getOriginalJson());
               item.put("signatureAndroid", purchase.getSignature());
-              item.put("developerPayload", purchase.getDeveloperPayload());
               items.put(item);
             }
             result.success(items.toString());
@@ -327,7 +324,7 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
 
     /*
      * buyItemByType
-     * arguments: type, accountId, developerId,  sku, oldSku, prorationMode
+     * arguments: type, obfuscatedAccountId, obfuscatedProfileId, sku, oldSku, prorationMode, purchaseToken
      */
     else if (call.method.equals("buyItemByType")) {
       if (billingClient == null || !billingClient.isReady()) {
@@ -336,32 +333,33 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
       }
 
       final String type = call.argument("type");
-      final String accountId = call.argument("accountId");
-      final String developerId = call.argument("developerId");
+      final String obfuscatedAccountId = call.argument("obfuscatedAccountId");
+      final String obfuscatedProfileId = call.argument("obfuscatedProfileId");
       final String sku = call.argument("sku");
       final String oldSku = call.argument("oldSku");
       final int prorationMode = call.argument("prorationMode");
+      final String purchaseToken = call.argument("purchaseToken");
 
       BillingFlowParams.Builder builder = BillingFlowParams.newBuilder();
 
       if (type.equals(BillingClient.SkuType.SUBS) && oldSku != null && !oldSku.isEmpty()) {
         // Subscription upgrade/downgrade
-        builder.setOldSku(oldSku);
+        builder.setOldSku(oldSku, purchaseToken);
       }
 
       if (type.equals(BillingClient.SkuType.SUBS) && oldSku != null && !oldSku.isEmpty()) {
         // Subscription upgrade/downgrade
         if (prorationMode != -1) {
-          builder.setOldSku(oldSku);
+          builder.setOldSku(oldSku, purchaseToken);
           if (prorationMode == BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_PRORATED_PRICE) {
             builder.setReplaceSkusProrationMode(BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_PRORATED_PRICE);
           } else if (prorationMode == BillingFlowParams.ProrationMode.IMMEDIATE_WITHOUT_PRORATION) {
             builder.setReplaceSkusProrationMode(BillingFlowParams.ProrationMode.IMMEDIATE_WITHOUT_PRORATION);
           } else {
-            builder.setOldSku(oldSku);
+            builder.setOldSku(oldSku, purchaseToken);
           }
         } else {
-          builder.setOldSku(oldSku);
+          builder.setOldSku(oldSku, purchaseToken);
         }
       }
 
@@ -382,11 +380,11 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
         return;
       }
 
-      if (accountId != null) {
-        builder.setAccountId(accountId);
+      if (obfuscatedAccountId != null) {
+        builder.setObfuscatedAccountId(obfuscatedAccountId);
       }
-      if (developerId != null) {
-        builder.setDeveloperId(developerId);
+      if (obfuscatedProfileId != null) {
+        builder.setObfuscatedProfileId(obfuscatedProfileId);
       }
 
       builder.setSkuDetails(selectedSku);
@@ -396,11 +394,10 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
 
     /*
      * acknowledgePurchase (For non-consumable purchases)
-     * arguments: token, developerPayload
+     * arguments: token
      */
     else if (call.method.equals("acknowledgePurchase")) {
       final String token = call.argument("token");
-      final String developerPayload = call.argument("developerPayload");
 
       if (billingClient == null || !billingClient.isReady()) {
         result.error(call.method, "IAP not prepared. Check if Google Play service is available.", "");
@@ -410,7 +407,6 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
       AcknowledgePurchaseParams acknowledgePurchaseParams =
           AcknowledgePurchaseParams.newBuilder()
               .setPurchaseToken(token)
-              .setDeveloperPayload(developerPayload)
               .build();
       billingClient.acknowledgePurchase(acknowledgePurchaseParams, new AcknowledgePurchaseResponseListener() {
         @Override
@@ -437,7 +433,7 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
 
     /*
      * consumeProduct (For consumable purchases)
-     * arguments: token, developerPayload
+     * arguments: token
      */
     else if (call.method.equals("consumeProduct")) {
       if (billingClient == null || !billingClient.isReady()) {
@@ -446,11 +442,9 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
       }
 
       final String token = call.argument("token");
-      final String developerPayload = call.argument("developerPayload");
 
       final ConsumeParams params = ConsumeParams.newBuilder()
           .setPurchaseToken(token)
-          .setDeveloperPayload(developerPayload)
           .build();
       billingClient.consumeAsync(params, new ConsumeResponseListener() {
         @Override
@@ -515,7 +509,6 @@ public class AndroidInappPurchasePlugin implements MethodCallHandler {
             item.put("autoRenewingAndroid", purchase.isAutoRenewing());
             item.put("isAcknowledgedAndroid", purchase.isAcknowledged());
             item.put("purchaseStateAndroid", purchase.getPurchaseState());
-            item.put("developerPayloadAndroid", purchase.getDeveloperPayload());
             item.put("originalJsonAndroid", purchase.getOriginalJson());