+++ lib/flutter_inapp_purchase.dart
@@ -1,13 +1,13 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:meta/meta.dart';
-import 'package:platform/platform.dart';
-import 'package:http/http.dart' as http;
 
 import 'package:flutter/services.dart';
+import 'package:http/http.dart' as http;
+import 'package:meta/meta.dart';
+import 'package:platform/platform.dart';
 
-import 'utils.dart';
 import 'modules.dart';
+import 'utils.dart';
 
 export 'modules.dart';
 
@@ -222,8 +222,9 @@ class FlutterInappPurchase {
   ///
   /// Identical to [requestSubscription] on `iOS`.
   Future requestPurchase(String sku, {
-    String developerIdAndroid,
-    String accountIdAndroid,
+    String obfuscatedAccountIdAndroid,
+    String obfuscatedProfileIdAndroid,
+    String purchaseTokenAndroid,
   }) async {
     if (_platform.isAndroid) {
       return await _channel.invokeMethod('buyItemByType', <String, dynamic>{
@@ -231,8 +232,9 @@ class FlutterInappPurchase {
         'sku': sku,
         'oldSku': null,
         'prorationMode': -1,
-        'developerId': developerIdAndroid,
-        'accountId': accountIdAndroid,
+        'obfuscatedAccountId': obfuscatedAccountIdAndroid,
+        'obfuscatedProfileId': obfuscatedProfileIdAndroid,
+        'purchaseToken': purchaseTokenAndroid,
       });
     } else if (_platform.isIOS) {
       return await _channel.invokeMethod(
@@ -254,8 +256,9 @@ class FlutterInappPurchase {
       {
         String oldSkuAndroid,
         int prorationModeAndroid,
-        String developerIdAndroid,
-        String accountIdAndroid,
+        String obfuscatedAccountIdAndroid,
+        String obfuscatedProfileIdAndroid,
+        String purchaseTokenAndroid,
       }) async {
     if (_platform.isAndroid) {
       return await _channel
@@ -264,8 +267,9 @@ class FlutterInappPurchase {
         'sku': sku,
         'oldSku': oldSkuAndroid,
         'prorationMode': prorationModeAndroid ?? -1,
-        'developerId': developerIdAndroid,
-        'accountId': accountIdAndroid,
+        'obfuscatedAccountId': obfuscatedAccountIdAndroid,
+        'obfuscatedProfileId': obfuscatedProfileIdAndroid,
+        'purchaseToken': purchaseTokenAndroid,
       });
     } else if (_platform.isIOS) {
       return await _channel.invokeMethod(
@@ -349,11 +353,10 @@ class FlutterInappPurchase {
   /// Acknowledge a purchase on `Android`.
   ///
   /// No effect on `iOS`, whose iap purchases are consumed at the time of purchase.
-  Future<String> acknowledgePurchaseAndroid(String token, { String developerPayload }) async {
+  Future<String> acknowledgePurchaseAndroid(String token) async {
     if (_platform.isAndroid) {
       String result = await _channel.invokeMethod('acknowledgePurchase', <String, dynamic>{
         'token': token,
-        'developerPayload': developerPayload,
       });
 
       return result;
@@ -377,12 +380,11 @@ class FlutterInappPurchase {
   ///        errorData[0] = E_ALREADY_OWNED;
   ///        errorData[1] = "You already own this item.";
   ///        break;
-  Future<String> consumePurchaseAndroid(String token, { String developerPayload }) async {
+  Future<String> consumePurchaseAndroid(String token) async {
     if (_platform.isAndroid) {
       String result =
           await _channel.invokeMethod('consumeProduct', <String, dynamic>{
         'token': token,
-        'developerPayload': developerPayload,
       });
       return result;
     } else if (_platform.isIOS) {
@@ -431,18 +433,16 @@ class FlutterInappPurchase {
   ///
   /// Call this after finalizing server-side validation of the reciept.
   Future<String> finishTransaction(PurchasedItem purchasedItem,
-    { String developerPayloadAndroid = '', bool isConsumable = false }) async {
+    { bool isConsumable = false }) async {
     if (_platform.isAndroid) {
       if (isConsumable) {
         String result = await _channel.invokeMethod('consumeProduct', <String, dynamic>{
           'token': purchasedItem.purchaseToken,
-          'developerPayload': developerPayloadAndroid,
         });
         return result;
       } else {
         String result = await _channel.invokeMethod('acknowledgePurchase', <String, dynamic>{
           'token': purchasedItem.purchaseToken,
-          'developerPayload': developerPayloadAndroid,
         });
         return result;
       }